packages/ywpi_hub/src/ywpi_hub/requirements.txt
aio_pika
aiochannel
grpcio-tools==1.64.0
pydantic
bson
